% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.variable.selection.R
\name{plot.variable.selection}
\alias{plot.variable.selection}
\title{Plot method for and object of class variable.selection}
\usage{
\method{plot}{variable.selection}(x, which = c(1L, 2L, 3L), col = "black",
  ...)
}
\arguments{
\item{x}{an object of class variable.selection as produced by \code{variable.selection()}.}

\item{which}{a numeric vector with the required plots. By default, all the possible options are plotted.}

\item{col}{color}

\item{...}{further arguments passed to method plot.default.}
}
\description{
On the basis of a variable.selection object, two/three plots are currently available:
(1) a plot showinh the variables selected in each method ;
(2) a plot showing the  frequency of the selection of each variable  across all methods;
and, if there is any ranker methods selected, (3) a plot showing the weight of each variable in each method. The square related to each variable-method will be darker if the  variable has a high weight in this method and it will be white if it has a low weight. The intensity of color shows the weight of the variable.
}
\examples{
library(VariableSelection)
 data(tlgs)

 ###########################################
 # Variable Selection with the deafault method (all methods with the default parameters):
 object <- variable.selection (input=names(tlgs)[-1],
 target=names(tlgs)[1], data=tlgs)
 plot(object)
}
\author{
Farideh Bagherzadeh-Khiabani.
}

