% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.variable.selection.R
\name{print.variable.selection}
\alias{print.variable.selection}
\title{Print method for and object of class variable.selection}
\usage{
\method{print}{variable.selection}(x, ...)
}
\arguments{
\item{x}{an object of class variable.selection as produced by \code{variable.selection()}.}

\item{...}{further arguments passed to method print.default.}
}
\description{
A short summary is printed with: the call to the variable.selection() function;
Then On the basis of a variable.selection object, two/three heading are currently available:
(1) "The selected variable(s)" shows the selected variables according to each method;
(2) "Weights for variables" emereges if there is any ranker methods selected,  showing the weight of each variable in each method.
and (3)  "Frequency of Selection" shows the frequency of the selection of each variable was selected across all methods.
}
\examples{
library(VariableSelection)
 data(tlgs)

 ###########################################
 # Variable Selection with the deafault method (all methods with the default parameters):
 object <- variable.selection (input=names(tlgs)[-1],
 target=names(tlgs)[1], data=tlgs)
 print(object)
}
\author{
Farideh Bagherzadeh-Khiabani.
}

